+++ src/connector.ts
@@ -631,6 +631,21 @@ export async function connect(
   } catch (e) {
     console.error(e);
   }
+  if (!['babashka', 'nbb', 'joyride', 'generic'].includes(connectSequence.projectType)) {
+    await nClient.session.info('clojure.core', 'map').catch((e) => {
+      void vscode.window
+        .showWarningMessage(
+          "Calva failed to perform a basic nREPL 'info' call. You need to start the REPL with cider-nrepl dependencies met",
+          'Show Calva Connect Docs'
+        )
+        .then((choice) => {
+          if (choice === 'Show Calva Connect Docs') {
+            void vscode.commands.executeCommand('simpleBrowser.show', 'https://calva.io/connect/');
+          }
+        });
+      console.error(`cider-nrepl dependencies not met: `, e);
+    });
+  }
   return true;
 }